refactor where Config lives
