+++ clarifai/utils/constants.py
@@ -1,3 +1,8 @@
+import os
+
 CLARIFAI_PAT_ENV_VAR = "CLARIFAI_PAT"
 CLARIFAI_SESSION_TOKEN_ENV_VAR = "CLARIFAI_SESSION_TOKEN"
 CLARIFAI_USER_ID_ENV_VAR = "CLARIFAI_USER_ID"
+
+
+DEFAULT_CONFIG = f'{os.environ["HOME"]}/.config/clarifai/config'
